+++ Bioinformatics/src/main/scala-2.11/utils/Implicits.scala
@@ -5,7 +5,7 @@ object Implicits {
   implicit class ExtendedInteger(n: Int) {
     import ExtendedInteger.calcBinomialCoefficient
 
-    def choose(k: Int): Int = {
+    def choose(k: Int): Long = {
       require(k >= 0 && k <= n)
       if (n - k < k) choose(n - k)
       else calcBinomialCoefficient(n, k)
@@ -14,9 +14,9 @@ object Implicits {
   }
 
   object ExtendedInteger {
-    def calcBinomialCoefficient(n: Int, k: Int): Int = {
-      if (k == 0) 1
-      else if (k == 1) n
+    def calcBinomialCoefficient(n: Int, k: Int): Long = {
+      if (k == 0) 1L
+      else if (k == 1) n.toLong
       else (n * calcBinomialCoefficient(n - 1, k - 1)) / k
     }
   }
